Added linear and angular dynamics for individual objects

+++ src/PhysicsEngine/Core/PhysicsEngine.hpp
@@ -7,7 +7,7 @@ using namespace std;
 class PhysicsEngine {
     static vector<PhysicsObject> physicsObjects;
 public:
-    static void update(float deltaTime);
+    static void update(unsigned int deltaTime);
     static PhysicsObject& createObject(sf::Vector3f position);
     static PhysicsObject& createObject();
